front/syn_metadata: move #[elidable_promote] synthesizer out of front/ast (Slice 2)
Relocates the elidable_promote / purefunction_promote cluster:
  PromoteArgsSelector, parse_elidable_promote_args,
  extract_elidable_promote_selector, synthesize_or_passthrough,
  synthesize_elidable_promote_pair.

The cluster is a pure syn::ItemFn -> [orig, wrapper] tree
transformation; no graph-builder state participates. parse.rs callsites
(extract_trait_impls / extract_inherent_impl_methods / build_graphs)
now reference crate::front::syn_metadata::synthesize_or_passthrough.

front::ast drops the local ItemFn import (now reached through the
helpers).
