+++ majit/majit-translate/src/front/ast.rs
@@ -4,7 +4,7 @@
 //! This module lowers syn AST nodes into FunctionGraph ops with proper data flow (Variable linking).
 
 use std::collections::HashMap;
-use syn::{Item, ItemFn};
+use syn::Item;
 
 use crate::ParsedInterpreter;
 use crate::flowspace::model::ConstValue;
@@ -512,340 +512,6 @@ fn collect_fields_and_returns(
             _ => {}
         }
     }
-}
-
-/// `promote_args` selector for `#[elidable_promote(...)]` synthesis.
-///
-/// Mirrors `rlib/jit.py:189-191` — RPython splits a literal string at
-/// `,` when the value is not the special `"all"` marker:
-///
-/// ```python
-/// if promote_args != 'all':
-///     args = [args[int(i)] for i in promote_args.split(",")]
-/// ```
-#[derive(Debug, Clone)]
-enum PromoteArgsSelector {
-    /// `jit.py:180` default `promote_args='all'` — every non-self
-    /// positional arg flows through `hint(..., promote=True)`.
-    All,
-    /// `jit.py:189-191` index list (`"0,2"` → `[0, 2]`).  Indices are
-    /// 0-based and point into the **positional arg list including
-    /// `self`** when the decorated function is a method (RPython's
-    /// `_get_args(func)` reads the raw co_varnames; pyre mirrors the
-    /// same convention by treating `self` as index 0 when present).
-    Indices(Vec<usize>),
-}
-
-/// `rlib/jit.py:180-191` — parse the literal attribute value attached
-/// to `#[elidable_promote(promote_args = "...")]`.  Bare
-/// `#[elidable_promote]` defaults to `"all"` per the upstream default
-/// argument at jit.py:180.
-fn parse_elidable_promote_args(attr: &syn::Attribute) -> syn::Result<PromoteArgsSelector> {
-    if matches!(attr.meta, syn::Meta::Path(_)) {
-        return Ok(PromoteArgsSelector::All);
-    }
-    let mut selector = None;
-    attr.parse_nested_meta(|meta| {
-        if meta.path.is_ident("promote_args") {
-            let value = meta.value()?;
-            let lit: syn::LitStr = value.parse()?;
-            selector = Some(if lit.value() == "all" {
-                PromoteArgsSelector::All
-            } else {
-                let mut indices = Vec::new();
-                for piece in lit.value().split(',') {
-                    indices.push(piece.trim().parse::<usize>().map_err(|err| {
-                        syn::Error::new(lit.span(), format!("promote_args: {err}"))
-                    })?);
-                }
-                PromoteArgsSelector::Indices(indices)
-            });
-            Ok(())
-        } else {
-            Err(meta.error("unsupported elidable_promote argument"))
-        }
-    })?;
-    Ok(selector.unwrap_or(PromoteArgsSelector::All))
-}
-
-/// Expand a `syn::ItemFn` into `[orig, wrapper]` when it carries
-/// `#[elidable_promote]` / `#[purefunction_promote]`, else return the
-/// single function unchanged.  Mirrors `rlib/jit.py:184-201`'s "module
-/// import installs two callables" semantics at every entry point that
-/// hands a `syn::ItemFn` to `build_function_graph*` — free functions
-/// (`build_graphs_from_items`), inherent / trait-impl methods
-/// (`parse::extract_inherent_impl_methods`,
-/// `parse::extract_trait_impls`), and trait default methods
-/// (`parse::extract_trait_impls`).  Centralising the expansion here
-/// keeps the decorator behaviour uniform across all four lowering
-/// surfaces.
-///
-/// `impl_type` carries the qualified type root (`"S"`, `"a::S"`) when
-/// the source item lives inside an `impl` block, so the synthesizer can
-/// emit a type-qualified tail call (`a::S::_orig_<name>_unlikely_name(
-/// self, args)`) that matches the impl-method registration path built
-/// by `parse::CallPath::for_impl_method` at `lib.rs:531-537`.  Free
-/// functions and trait-default methods (which have no concrete `Self`
-/// type at synthesis time) pass `None` and fall back to the bare-path
-/// tail call.
-pub fn synthesize_or_passthrough(fake_fn: ItemFn, impl_type: Option<&str>) -> Vec<ItemFn> {
-    match extract_elidable_promote_selector(&fake_fn.attrs) {
-        Some(selector) => {
-            let (orig, wrapper) = synthesize_elidable_promote_pair(&fake_fn, &selector, impl_type);
-            vec![orig, wrapper]
-        }
-        None => vec![fake_fn],
-    }
-}
-
-/// Locate `#[elidable_promote]` (or its deprecated `#[purefunction_promote]`
-/// alias from `rlib/jit.py:203-205`) and return its parsed
-/// `promote_args` selector, or `None` if neither attribute is present.
-///
-/// `rlib/jit.py:189-191` `args[int(i)]` propagates `ValueError` /
-/// `IndexError` to the caller — the decorator does not silently drop
-/// malformed input.  Pyre mirrors that fail-loud behaviour here: a
-/// malformed `promote_args = "..."` literal panics with the
-/// underlying `syn::Error` rather than falling through.
-fn extract_elidable_promote_selector(attrs: &[syn::Attribute]) -> Option<PromoteArgsSelector> {
-    for attr in attrs {
-        if let Some(segment) = attr.path().segments.last() {
-            let name = segment.ident.to_string();
-            if name == "elidable_promote" || name == "purefunction_promote" {
-                return Some(parse_elidable_promote_args(attr).unwrap_or_else(|err| {
-                    panic!("#[{name}(...)]: {err}");
-                }));
-            }
-        }
-    }
-    None
-}
-
-/// Synthesize the wrapper / original function pair from a single
-/// `#[elidable_promote] fn foo(...)` source item.
-///
-/// Line-by-line port of `rlib/jit.py:184-201`:
-///
-/// ```python
-/// def decorator(func):
-///     elidable(func)                                # func._elidable_function_ = True
-///     args = _get_args(func)
-///     code = ["def f(%s):\n" % (argstring,)]
-///     if promote_args != 'all':
-///         args = [args[int(i)] for i in promote_args.split(",")]
-///     for arg in args:
-///         code.append("    %s = hint(%s, promote=True, promote_string=True)\n" % …)
-///     code.append("    return _orig_func_unlikely_name(%s)\n" % …)
-///     d = {"_orig_func_unlikely_name": func, "hint": hint}
-///     exec py.code.Source("\n".join(code)).compile() in d
-///     result = d["f"]
-///     result.__name__ = func.__name__ + "_promote"
-///     return result
-/// ```
-///
-/// Pyre's syn-tree adaptation:
-///
-///   * RPython's `_orig_func_unlikely_name` closure capture becomes a
-///     module-scope sibling `fn _orig_<name>_unlikely_name` carrying
-///     the original body and the `#[elidable]` attribute (so
-///     `collect_jit_hints` registers it via `mark_elidable`).
-///   * RPython's `exec compile` synthesis of `def f(...)` becomes a
-///     `syn::Block` whose stmts are
-///     `let <arg> = hint_promote_or_string(<arg>);` for each selected
-///     arg, followed by the tail call.
-///   * The user-facing name keeps its identity (matching pyre's
-///     proc-macro naming convention `_orig_<name>_unlikely_name` at
-///     `majit-macros::elidable_promote`); RPython's `__name__ + "_promote"`
-///     rename is the symmetric choice and equivalently produces two
-///     distinct callable identifiers.
-fn synthesize_elidable_promote_pair(
-    func: &ItemFn,
-    selector: &PromoteArgsSelector,
-    impl_type: Option<&str>,
-) -> (ItemFn, ItemFn) {
-    use quote::format_ident;
-    // jit.py:186 args = _get_args(func) — positional names, self included.
-    let arg_names: Vec<syn::Ident> = func
-        .sig
-        .inputs
-        .iter()
-        .map(|arg| match arg {
-            syn::FnArg::Typed(pt) => match &*pt.pat {
-                syn::Pat::Ident(pi) => pi.ident.clone(),
-                _ => panic!(
-                    "#[elidable_promote] on `fn {}`: unsupported binder \
-                     pattern in arg position — RPython `_get_args(func)` \
-                     (jit.py:172-178) reads positional names off \
-                     `co_varnames`, which never include destructured \
-                     binders.  Rewrite the parameter as a plain `name: \
-                     Ty` instead.",
-                    func.sig.ident
-                ),
-            },
-            // `&self` / `&mut self` map to a positional name "self" so
-            // index 0 in `Indices` continues to address the receiver as
-            // RPython does (`_get_args(func)` reads co_varnames raw).
-            syn::FnArg::Receiver(_) => format_ident!("self"),
-        })
-        .collect();
-
-    let orig_ident = format_ident!("_orig_{}_unlikely_name", func.sig.ident);
-
-    // jit.py:184-185 — original keeps the body, gains `_elidable_function_`.
-    let orig_attrs: Vec<syn::Attribute> = func
-        .attrs
-        .iter()
-        .filter(|a| {
-            let name = a
-                .path()
-                .segments
-                .last()
-                .map(|s| s.ident.to_string())
-                .unwrap_or_default();
-            name != "elidable_promote" && name != "purefunction_promote"
-        })
-        .cloned()
-        .collect();
-    let elidable_attr: syn::Attribute = syn::parse_quote!(#[elidable]);
-    let mut orig_fn = func.clone();
-    orig_fn.attrs = orig_attrs;
-    orig_fn.attrs.push(elidable_attr);
-    orig_fn.sig.ident = orig_ident.clone();
-
-    // jit.py:189-191 — promote_args=='all' → all indices; else parsed list.
-    // RPython `args[int(i)]` raises `IndexError` on out-of-range
-    // indices; pyre fails loudly here for the same reason.
-    let promote_indices: Vec<usize> = match selector {
-        PromoteArgsSelector::All => (0..arg_names.len()).collect(),
-        PromoteArgsSelector::Indices(ix) => {
-            for &i in ix {
-                if i >= arg_names.len() {
-                    panic!(
-                        "#[elidable_promote(promote_args = ...)] on `fn {}`: \
-                         index {} is out of range for a {}-arg function \
-                         (jit.py:191 `args[int(i)]` would IndexError)",
-                        func.sig.ident,
-                        i,
-                        arg_names.len()
-                    );
-                }
-            }
-            ix.clone()
-        }
-    };
-    // jit.py:191-194 — `for arg in args: hint(arg, promote=True,
-    // promote_string=True)`.  RPython emits a *single* dual-flag hint
-    // per arg and defers the dispatch to
-    // `rpython/jit/codewriter/jtransform.py:599-606`, which only
-    // matches `concretetype == lltype.Ptr(rstr.STR)`:
-    //
-    //   * `Ptr(rstr.STR)` → keep `promote_string`, delete `promote`
-    //     → emit `str_guard_value` (jit.py:615-631).
-    //   * everything else, **including `Ptr(rstr.UNICODE)`** → delete
-    //     `promote_string`, keep `promote` → emit
-    //     `<kind>_guard_value` with `kind = getkind(concretetype)`,
-    //     which is `"ref"` for any GC pointer
-    //     (`rpython/jit/metainterp/history.py:64-67`).  No
-    //     `unicode_guard_value` opname exists in RPython; the
-    //     unicode-specific `str_guard_value/OS_UNIEQ_NONNULL` shape
-    //     is only reachable via the dedicated `promote_unicode`
-    //     hint (`jit_codewriter/jtransform.py:632-648`,
-    //     `rpython/rlib/jit.py:130-131`), which `elidable_promote`
-    //     never emits.
-    //
-    // Pyre mirrors that shape line-by-line: every promoted arg gets
-    // `hint_promote_or_string`, and the `PromoteOrString` arm in
-    // `jit_codewriter/jtransform.rs` falls through to the plain
-    // `<kind>_guard_value` family because pyre lacks a
-    // `Ptr(rstr.STR)`-equivalent GC layout — `del hints['promote_string']`
-    // is the upstream `else` branch at `jtransform.py:603-606`.
-    //
-    // `args` includes `self` since `_get_args(func)` reads
-    // `co_varnames` raw.  Rust forbids re-binding the `self` keyword,
-    // so the receiver is routed through a fresh `__self_promoted`
-    // local; non-receiver args keep RPython's shadow pattern.
-    let promote_self = promote_indices.iter().any(|&i| arg_names[i] == "self");
-    let promote_stmts: Vec<syn::Stmt> = promote_indices
-        .iter()
-        .map(|&i| {
-            let id = &arg_names[i];
-            if id == "self" {
-                syn::parse_quote!(let __self_promoted = hint_promote_or_string(self);)
-            } else {
-                syn::parse_quote!(let #id = hint_promote_or_string(#id);)
-            }
-        })
-        .collect();
-
-    // jit.py:195 — return _orig_func_unlikely_name(args).  When `self` was
-    // promoted above, the tail call uses the promoted local; otherwise it
-    // threads the unmodified receiver through.
-    let call_args = arg_names.iter().map(|id| -> syn::Expr {
-        if id == "self" {
-            if promote_self {
-                syn::parse_quote!(__self_promoted)
-            } else {
-                syn::parse_quote!(self)
-            }
-        } else {
-            syn::parse_quote!(#id)
-        }
-    });
-    // jit.py:195-197 — `_orig_func_unlikely_name` is bound in the
-    // wrapper's `exec` namespace (`d = {"_orig_func_unlikely_name":
-    // func, ...}`), so the wrapper invokes it as a closure-local name.
-    // Pyre's lowering surface does not have an analogue closure scope;
-    // for free fns the orig is registered as a bare-path function (so a
-    // bare-path call resolves), but for impl methods the orig is
-    // registered under `<ImplType>::_orig_<name>_unlikely_name` via
-    // `parse::CallPath::for_impl_method` at `lib.rs:531-537`.  Emit a
-    // type-qualified tail call in that case so the wrapper's IR `Call`
-    // target matches the registration path and the elidable flag bound
-    // to the orig graph is visible at the call-site.
-    let tail_call: syn::Expr = match impl_type {
-        Some(ty_str) => {
-            let ty_path: syn::Path = syn::parse_str(ty_str).unwrap_or_else(|err| {
-                panic!(
-                    "synthesize_elidable_promote_pair: failed to parse impl type `{ty_str}`: {err}"
-                )
-            });
-            syn::parse_quote!(#ty_path::#orig_ident(#(#call_args),*))
-        }
-        None => syn::parse_quote!(#orig_ident(#(#call_args),*)),
-    };
-    let tail_stmt = syn::Stmt::Expr(tail_call, None);
-
-    // jit.py:198-201 — wrapper is the user-facing decorated name with
-    // the promote+forward body; `#[elidable_promote]` is stripped so
-    // `collect_jit_hints` does not register the wrapper itself as
-    // elidable (the binary flag belongs on the original alone, per
-    // jit.py:185 `elidable(func)`).
-    let wrapper_attrs: Vec<syn::Attribute> = func
-        .attrs
-        .iter()
-        .filter(|a| {
-            let name = a
-                .path()
-                .segments
-                .last()
-                .map(|s| s.ident.to_string())
-                .unwrap_or_default();
-            name != "elidable_promote" && name != "purefunction_promote"
-        })
-        .cloned()
-        .collect();
-    let mut wrapper_block = syn::Block {
-        brace_token: Default::default(),
-        stmts: Vec::new(),
-    };
-    wrapper_block.stmts.extend(promote_stmts);
-    wrapper_block.stmts.push(tail_stmt);
-
-    let mut wrapper_fn = func.clone();
-    wrapper_fn.attrs = wrapper_attrs;
-    wrapper_fn.block = Box::new(wrapper_block);
-
-    (orig_fn, wrapper_fn)
 }
 
 /// Walk the statements of a fn body and register the return types of
